requirements.txt
streamlit
ultralytics
numpy
pillow
